README.md: Remove most picture sizes from HTML table
Some CSS rules set "max-width: 100%;", so the 1500px width we keep for
the first picture is not actually respected. However, it seems that this
width indication is taken into account to compute the width of the
column with respect to the width of the second column, so we want to
keep it to something rather high to get wider pictures than what is
computed by default: we leave the width untouched for the Tetragon eBee.

The width of the first row is enough to set the width of the column for
the full table, let's get rid of the other width indications.

Signed-off-by: Quentin Monnet <quentin@isovalent.com>